# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, plugin.
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/tools:
    get:
      tags: [Tools]
      summary: List LLM tools
      operationId: listTools
      x-astrbot-scope: tool
      parameters:
        - name: origin
          in: query
          schema:
            type: string
            enum: [builtin, plugin, mcp]
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/tools/{tool_id}/enabled:
    patch:
      tags: [Tools]
      summary: Enable or disable an LLM tool
      operationId: setToolEnabled
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ToolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledPatch"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/tools/{tool_id}/permission:
    patch:
      tags: [Tools]
      summary: Update an LLM tool permission
      operationId: setToolPermission
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ToolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolPermissionPatch"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/mcp/servers:
    get:
      tags: [MCP]
      summary: List MCP servers
      operationId: listMcpServers
      x-astrbot-scope: tool
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [MCP]
      summary: Add an MCP server
      operationId: createMcpServer
      x-astrbot-scope: tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpServerConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/mcp/servers/by-name:
    put:
      tags: [MCP]
      summary: Update an MCP server by body name
      operationId: updateMcpServerByName
      x-astrbot-scope: tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_name]
              properties:
                server_name:
                  type: string
                config:
                  $ref: "#/components/schemas/DynamicConfig"
                enabled:
                  type: boolean
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [MCP]
      summary: Delete an MCP server by query name
      operationId: deleteMcpServerByName
      x-astrbot-scope: tool
      parameters:
        - name: server_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/mcp/servers/enabled:
    patch:
      tags: [MCP]
      summary: Enable or disable an MCP server by body name
      operationId: setMcpServerEnabledByName
      x-astrbot-scope: tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_name, enabled]
              properties:
                server_name:
                  type: string
                enabled:
                  type: boolean
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/mcp/servers/test:
    post:
      tags: [MCP]
      summary: Test an MCP server connection by body name
      operationId: testMcpServerByName
      x-astrbot-scope: tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_name]
              properties:
                server_name:
                  type: string
                mcp_server_config:
                  $ref: "#/components/schemas/DynamicConfig"
                config:
                  $ref: "#/components/schemas/DynamicConfig"
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/mcp/servers/{server_name}:
    put:
      tags: [MCP]
      summary: Update an MCP server
      operationId: updateMcpServer
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ServerName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpServerConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [MCP]
      summary: Delete an MCP server
      operationId: deleteMcpServer
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ServerName"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/mcp/servers/{server_name}/enabled:
    patch:
      tags: [MCP]
      summary: Enable or disable an MCP server
      operationId: setMcpServerEnabled
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ServerName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledPatch"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/mcp/servers/{server_name}/test:
    post:
      tags: [MCP]
      summary: Test an MCP server connection
      operationId: testMcpServer
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ServerName"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mcp_server_config:
                  type: object
                  additionalProperties: true
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/mcp/providers/modelscope/sync:
    post:
      tags: [MCP]
      summary: Sync MCP servers from ModelScope
      operationId: syncModelScopeMcpServers
      x-astrbot-scope: tool
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelScopeSyncRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    ServerName:
      name: server_name
      in: path
      required: true
      schema:
        type: string
...
    ToolId:
      name: tool_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    EnabledPatch:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
      additionalProperties: false
    ToolPermissionPatch:
      type: object
      required: [permission]
      properties:
        permission:
          type: string
          enum: [admin, member]
      additionalProperties: false
...
    McpServerConfig:
      type: object
      required: [name]
      properties:
        name:
          type: string
        enabled:
          type: boolean
          default: true
        transport:
          type: string
          enum: [stdio, sse, streamable_http]
        command:
          type: string
        args:
          type: array
          items:
            type: string
        url:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        timeout:
          type: integer
      additionalProperties: true
    ModelScopeSyncRequest:
      type: object
      properties:
        access_token:
          type: string
      additionalProperties: false
...